refactor BSD-only values out of POSIXFileSystem
Also adjust the visibility of the members of the NativeFileReader
subclass, and remove some of its unused constants.
