+++ GVFS/GVFS.Platform.POSIX/POSIXPlatform.cs
@@ -21,7 +21,7 @@ namespace GVFS.Platform.POSIX
         }
 
         public override IGitInstallation GitInstallation { get; } = new POSIXGitInstallation();
-        public override IPlatformFileSystem FileSystem { get; } = new POSIXFileSystem();
+
         public override void ConfigureVisualStudio(string gitBinPath, ITracer tracer)
         {
         }
